fix: Move rescheduleJobs into Utils
