+++ src/DIRAC/WorkloadManagementSystem/Utilities/JobStatusUtility.py
@@ -9,9 +9,10 @@ from DIRAC import S_ERROR, S_OK, gLogger
 from DIRAC.Core.Utilities import TimeUtilities
 from DIRAC.Core.Utilities.ObjectLoader import ObjectLoader
 from DIRAC.WorkloadManagementSystem.Client import JobStatus
-from DIRAC.WorkloadManagementSystem.DB.JobLoggingDB import JobLoggingDB
-from DIRAC.WorkloadManagementSystem.DB.JobDB import JobDB
-from DIRAC.WorkloadManagementSystem.DB.TaskQueueDB import TaskQueueDB
+
+if TYPE_CHECKING:
+    from DIRAC.WorkloadManagementSystem.DB.JobLoggingDB import JobLoggingDB
+    from DIRAC.WorkloadManagementSystem.DB.JobDB import JobDB
 
 
 class JobStatusUtility:
@@ -242,35 +243,3 @@ def getNewStatus(
         minor = sDict.get("MinorStatus", minor)
         application = sDict.get("ApplicationStatus", application)
     return S_OK((status, minor, application))
-
-
-def rescheduleJobs(jobIDs: list[int], source: str = "") -> dict:
-    """Utility to reschedule jobs (not atomic, nor bulk)
-    Requires direct access to the JobDB and TaskQueueDB
-
-    :param jobIDs: list of jobIDs
-    :param source: source of the reschedule
-    :return: S_OK/S_ERROR
-    :rtype: dict
-
-    """
-
-    failedJobs = []
-
-    for jobID in jobIDs:
-        result = JobDB().rescheduleJob(jobID)
-        if not result["OK"]:
-            failedJobs.append(jobID)
-            continue
-        TaskQueueDB().deleteJob(jobID)
-        JobLoggingDB().addLoggingRecord(
-            result["JobID"],
-            status=result["Status"],
-            minorStatus=result["MinorStatus"],
-            applicationStatus="Unknown",
-            source=source,
-        )
-
-    if failedJobs:
-        return S_ERROR(f"Failed to reschedule jobs {failedJobs}")
-    return S_OK()
